+++ js/main.js
@@ -1,4 +1,4 @@
-import './generate-comment.js';
-import './generate-photos.js';
+import './util.js';
+import './data.js';
